+++ src/Linear/Matrix.hs
@@ -187,6 +187,7 @@ fromQuaternion (Quaternion w (V3 x y z)) =
   where x2 = x * x
         y2 = y * y
         z2 = z * z
+{-# INLINABLE fromQuaternion #-}
 
 -- | Build a transformation matrix from a rotation matrix and a
 -- translation vector.
@@ -194,11 +195,13 @@ mkTransformationMat :: Num a => M33 a -> V3 a -> M44 a
 mkTransformationMat (V3 r1 r2 r3) (V3 tx ty tz) =
   V4 (snoc3 r1 tx) (snoc3 r2 ty) (snoc3 r3 tz) (V4 0 0 0 1)
   where snoc3 (V3 x y z) = V4 x y z
+{-# INLINABLE mkTransformationMat #-}
 
 -- |Build a transformation matrix from a rotation expressed as a
 -- 'Quaternion' and a translation vector.
 mkTransformation :: Num a => Quaternion a -> V3 a -> M44 a
 mkTransformation = mkTransformationMat . fromQuaternion
+{-# INLINABLE mkTransformation #-}
 
 -- | Convert from a 4x3 matrix to a 4x4 matrix, extending it with the @[ 0 0 0 1 ]@ column vector
 m43_to_m44 :: Num a => M43 a -> M44 a